Fix SDK review issues: security, types, threading, packaging
Critical fixes:
- URL-encode IP parameter in Python, Java, Ruby SDKs (prevents injection)
- Fix Go SDK base URL: api.verifip.io → api.verifip.com (mismatch)
- Fix Go SDK FraudScore: float64 → int (matches server contract)
- Fix TypeScript timer leak: clearTimeout before return on success

High fixes:
- Python: guard from_dict() against None/non-dict input
- Python: cast retry_after to float (prevents TypeError on string)
- Python: use timezone.utc instead of UTC (Python 3.10 compat)
- Python: case-insensitive header lookup for rate limit parsing
- Python: handle ValueError on malformed rate limit headers
- Ruby: URI.encode_uri_component for IP parameter
- Ruby: thread-safe rate_limit_info with Mutex
- Go: check ctx.Err() after Do() in retry loop
- Fix root .gitignore: /verifip (binary only, not sdk/*/verifip/)

Packaging:
- Add MIT LICENSE to all 6 SDKs
- Add .gitignore to PHP, Java, Ruby, TypeScript SDKs

+++ models.go
@@ -6,7 +6,7 @@ import "time"
 type CheckResponse struct {
 	RequestID      string         `json:"request_id"`
 	IP             string         `json:"ip"`
-	FraudScore     float64        `json:"fraud_score"`
+	FraudScore     int            `json:"fraud_score"`
 	IsProxy        bool           `json:"is_proxy"`
 	IsVPN          bool           `json:"is_vpn"`
 	IsTor          bool           `json:"is_tor"`
